rabbitmq/requirements.txt
pika>=1.3.2
python-dotenv>=1.2.1
requests>=2.32.5

